aegis/backend/requirements.txt
# ──────────────────────────────────────────────────────────────
# AEGIS Backend — requirements.txt
# ──────────────────────────────────────────────────────────────

# ─── Core Framework ───
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-multipart>=0.0.6

# ─── Async & Networking ───
httpx>=0.26.0,<1.0.0
aiofiles>=23.2.0

# ─── Redis ───
redis>=5.0.0,<6.0.0

# ─── Git Operations ───
gitpython>=3.1.40,<4.0.0

# ─── Code Parsing ───
tree-sitter>=0.21.0,<1.0.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0

# ─── Machine Learning / AI Detection ───
scikit-learn>=1.4.0,<2.0.0
numpy>=1.26.0,<2.0.0
xgboost>=2.0.0,<3.0.0
optuna>=3.5.0,<4.0.0
joblib>=1.3.0,<2.0.0
pandas>=2.1.0,<3.0.0

# ─── Security Scanning ───
# semgrep is installed separately via pip install semgrep

# ─── Cryptography ───
cryptography>=42.0.0,<43.0.0

# ─── Azure (optional) ───
openai>=1.10.0,<2.0.0

# ─── Logging ───
structlog>=24.1.0,<25.0.0

# ─── Testing ───
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0
